+++ tests/cases/static_nglob/main.sh
@@ -38,11 +38,7 @@ watch_update("ch-2-theory/sec-2-2-original.txt")
 run()
 wait()
 graph("current_graph_02")
-join()
 EOD
-
-# Wait for background processes, if any.
-wait $(jobs -p)
 
 # Check files that are expected to be present and/or missing.
 [[ -f plan.py ]] || exit -1
@@ -64,3 +60,19 @@ wait $(jobs -p)
 [[ -f ch-3-conclusions/sec-3-2-outlook.md ]] || exit -1
 [[ -f ch-3-conclusions/ch-3-compiled.md ]] || exit -1
 [[ -f book.md ]] || exit -1
+
+# Test cleanup
+cleanup ch-3-conclusions/sec-3-1-summary.txt
+[[ -f ch-3-conclusions/sec-3-1-summary.txt ]] || exit -1
+[[ ! -f ch-3-conclusions/sec-3-1-summary.md ]] || exit -1
+[[ -f ch-3-conclusions/sec-3-2-outlook.md ]] || exit -1
+[[ ! -f ch-3-conclusions/ch-3-compiled.md ]] || exit -1
+[[ ! -f book.md ]] || exit -1
+
+python3 - << EOD
+from stepup.core.interact import *
+join()
+EOD
+
+# Wait for background processes, if any.
+wait $(jobs -p)